Refactor Call(AndAssign)Instruction in order to remove duplicate code
